sticker_maker_app/requirements.txt
flask
google-genai
rembg
Pillow
python-dotenv
opencv-python-headless
numpy
waitress
